refactor(application): extract profileApplicationService
